Remove mogenius from Deployment guide
Hi there, mogenius doesn't have free cloud resources anymore. It is possible to deploy any container app on the free plan but it requires to run a local Kubernetes. I suppose this is not a use case for the Whatsapp bot so I suggest to remove mogenius from the deployment guide.
Cheers,
Jan